requirements.txt
Flask
flask-login
authlib
requests
python-dotenv
Flask-SQLAlchemy
